requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
idna==3.1
paho-mqtt==1.5.1
python-engineio==3.8.2.post1
python-socketio==4.2.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
websocket-client==0.57.0
PyYAML==5.4.1
